refactor: add modules to style
