+++ rs-react-app/src/components/SearchForm/SearchForm.module.css
@@ -3,18 +3,18 @@
   justify-content: center;
   gap: 20px;
   margin-bottom: 20px;
-  .form-title {
+  .formTitle {
     padding-top: 7px;
     color: #646cff;
     font-weight: 500;
   }
-  .form-input {
+  .formInput {
     border-radius: 10px;
     padding: 0.6em;
     border: 1px solid #646cff;
     transition: border-color 0.25s;
   }
-  .form-input:focus {
+  .formInput:focus {
     border-color: #535bf2;
     outline: none;
   }
