+++ rs-react-app/src/components/searchResult/SearchResult.module.css
@@ -1,10 +1,10 @@
-.search-result {
+.searchResult {
   max-width: 1440px;
   padding: 0 60px;
   margin: 0 auto;
 }
 
-.list-items {
+.listItems {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
   gap: 10px;
