+++ rs-react-app/src/components/searchResult/card/Card.module.css
